split up handler test file
